Toggle zoom modes while scrolling
So space or a down arrow (or equivalent) on a best fit will first
go to width fit and space on the bottom of a non best fit will go
to best fit mode.

This does not take into account double pages, but I guess I will
leave this for when I have such comics to read.